[build] move default ThreadPool implementation to platform/default
