+++ platform/glfw/main.cpp
@@ -4,7 +4,7 @@
 #include <mbgl/util/default_styles.hpp>
 #include <mbgl/platform/log.hpp>
 #include <mbgl/platform/platform.hpp>
-#include <mbgl/platform/default/thread_pool.hpp>
+#include <mbgl/util/default_thread_pool.hpp>
 #include <mbgl/storage/default_file_source.hpp>
 
 #include <signal.h>
