+++ platform/macos/src/MGLMapView.mm
@@ -22,7 +22,7 @@
 #import <mbgl/annotation/annotation.hpp>
 #import <mbgl/map/camera.hpp>
 #import <mbgl/storage/reachability.h>
-#import <mbgl/platform/default/thread_pool.hpp>
+#import <mbgl/util/default_thread_pool.hpp>
 #import <mbgl/gl/extension.hpp>
 #import <mbgl/gl/context.hpp>
 #import <mbgl/map/backend.hpp>
