+++ test/actor/actor.test.cpp
@@ -1,5 +1,5 @@
 #include <mbgl/actor/actor.hpp>
-#include <mbgl/platform/default/thread_pool.hpp>
+#include <mbgl/util/default_thread_pool.hpp>
 
 #include <mbgl/test/util.hpp>
