+++ test/api/custom_layer.test.cpp
@@ -4,7 +4,7 @@
 #include <mbgl/map/map.hpp>
 #include <mbgl/platform/default/headless_backend.hpp>
 #include <mbgl/platform/default/offscreen_view.hpp>
-#include <mbgl/platform/default/thread_pool.hpp>
+#include <mbgl/util/default_thread_pool.hpp>
 #include <mbgl/storage/default_file_source.hpp>
 #include <mbgl/style/layers/custom_layer.hpp>
 #include <mbgl/style/layers/fill_layer.hpp>
